moved utils functions from v4 pcd package to a utils file
