+++ packages/pcd/semaphore-identity-v4/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./SemaphoreIdentityV4PCD";
 export * from "./SemaphoreIdentityV4PCDPackage";
+export * from "./utils";
